requirements.txt
python-dateutil~=2.8.1
fastapi~=0.63.0
boto3~=1.17.22
uvicorn~=0.13.4
mangum~=0.10.0